+++ apps/example/package.json
@@ -7,6 +7,7 @@
     "tsc": "tsc --noEmit",
     "android": "react-native run-android",
     "ios": "react-native run-ios",
+    "web": "IS_WEB_BUILD=true react-native start",
     "start": "react-native start",
     "pod:install:ios": "pod install --project-directory=ios",
     "pod:install:macos": "pod install --project-directory=macos",
@@ -19,18 +20,21 @@
     "@callstack/react-native-visionos": "^0.74.0",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/stack": "^6.4.0",
-    "@react-three/fiber": "^8.17.6",
+    "@react-three/fiber": "^9.1.2",
     "@shopify/react-native-skia": "2.0.0",
     "@tensorflow/tfjs": "^4.22.0",
     "@tensorflow/tfjs-backend-webgpu": "^4.22.0",
     "@tensorflow/tfjs-vis": "^1.5.1",
+    "@types/react-dom": "^19.1.5",
     "fast-text-encoding": "^1.0.6",
     "react": "19.0.0",
+    "react-dom": "19.0.0",
     "react-native": "0.79.2",
     "react-native-gesture-handler": "^2.17.1",
     "react-native-macos": "^0.78.3",
     "react-native-reanimated": "^3.12.1",
-    "react-native-safe-area-context": "^5.4.0",
+    "react-native-safe-area-context": "^5.4.1",
+    "react-native-web": "^0.20.0",
     "react-native-wgpu": "*",
     "teapot": "^1.0.0",
     "three": "0.172.0",
