+++ apps/example/src/MNISTInference/MNISTInference.tsx
@@ -1,7 +1,13 @@
 import React, { useCallback, useEffect, useMemo, useRef } from "react";
-import { Button, Dimensions, Platform, StyleSheet, View } from "react-native";
-import type { SkImage, SkSurface } from "@shopify/react-native-skia";
+import { Button, Dimensions, StyleSheet, View } from "react-native";
+import type {
+  SkImage,
+  SkPaint,
+  SkPath,
+  SkSurface,
+} from "@shopify/react-native-skia";
 import {
+  useFont,
   Canvas,
   Fill,
   Skia,
@@ -9,7 +15,6 @@ import {
   Path,
   ColorType,
   AlphaType,
-  matchFont,
   Text,
   Image,
   FilterMode,
@@ -21,40 +26,54 @@ import { useDevice } from "react-native-wgpu";
 import type { Network } from "./Lib";
 import { createDemo, centerData, SIZE } from "./Lib";
 
-const { width } = Dimensions.get("window");
-
-const fontFamily = Platform.select({ ios: "Helvetica", default: "serif" });
-const fontStyle = {
-  fontFamily,
-  fontSize: 200,
-};
-const font = matchFont(fontStyle);
-
-const paint = Skia.Paint();
-paint.setColor(Skia.Color("black"));
-paint.setStyle(PaintStyle.Stroke);
-paint.setStrokeWidth(1);
-
-const grid = Skia.Path.Make();
-const cellSize = width / SIZE;
-
-grid.moveTo(0, 0);
-
-// Draw vertical lines
-for (let i = 0; i <= SIZE; i++) {
-  grid.moveTo(i * cellSize, 0);
-  grid.lineTo(i * cellSize, width);
-}
-
-// Draw horizontal lines
-for (let i = 0; i <= SIZE; i++) {
-  grid.moveTo(0, i * cellSize);
-  grid.lineTo(width, i * cellSize);
-}
-
-const f = 1 / cellSize;
-
 export function MNISTInference() {
+  const skiaConstants = useRef<null | {
+    paint: SkPaint;
+    grid: SkPath;
+    width: number;
+    f: number;
+  }>(null);
+
+  const font = useFont(require("../assets/helvetica.ttf"));
+
+  // Lazy initialize skia derived constants
+  if (!skiaConstants.current) {
+    const { width } = Dimensions.get("window");
+
+    const paint = Skia.Paint();
+    paint.setColor(Skia.Color("black"));
+    paint.setStyle(PaintStyle.Stroke);
+    paint.setStrokeWidth(1);
+
+    const grid = Skia.Path.Make();
+    const cellSize = width / SIZE;
+
+    grid.moveTo(0, 0);
+
+    // Draw vertical lines
+    for (let i = 0; i <= SIZE; i++) {
+      grid.moveTo(i * cellSize, 0);
+      grid.lineTo(i * cellSize, width);
+    }
+
+    // Draw horizontal lines
+    for (let i = 0; i <= SIZE; i++) {
+      grid.moveTo(0, i * cellSize);
+      grid.lineTo(width, i * cellSize);
+    }
+
+    const f = 1 / cellSize;
+
+    skiaConstants.current = {
+      f,
+      paint,
+      grid,
+      width,
+    };
+  }
+
+  const { f, paint, grid, width } = skiaConstants.current;
+
   const { device } = useDevice();
   const network = useRef<Network>();
   const text = useSharedValue("");
@@ -84,21 +103,26 @@ export function MNISTInference() {
         if (surface.value) {
           const canvas = surface.value.getCanvas();
           canvas.drawPath(path.value, paint);
+          surface.value.flush();
           image.value = surface.value!.makeImageSnapshot();
           const pixels = image.value.readPixels(0, 0, {
             width: SIZE,
             height: SIZE,
-            alphaType: AlphaType.Opaque,
-            colorType: ColorType.Alpha_8,
+            colorType: ColorType.RGBA_8888,
+            alphaType: AlphaType.Unpremul,
           });
+
+          const gray = new Uint8Array(SIZE * SIZE);
+          for (let i = 0; i < SIZE * SIZE; i++) {
+            gray[i] = pixels![i * 4];
+          }
+
           runOnJS(runInference)(
-            centerData(pixels as Uint8Array).map(
-              (x) => (x / 255) * 3.24 - 0.42,
-            ),
+            centerData(gray).map((x) => (x / 255) * 3.24 - 0.42),
           );
         }
       });
-  }, [path, runInference, surface, image]);
+  }, [path, runInference, surface, image, f, paint]);
 
   useEffect(() => {
     (async () => {
@@ -111,6 +135,11 @@ export function MNISTInference() {
       })();
     })();
   }, [device, network, surface]);
+
+  if (!font) {
+    return null;
+  }
+
   return (
     <View style={style.container}>
       <Button
@@ -123,7 +152,7 @@ export function MNISTInference() {
         title="Reset"
       />
       <GestureDetector gesture={gesture}>
-        <Canvas style={style.canvas}>
+        <Canvas style={{ width, height: width * 2 }}>
           <Fill color="rgb(239, 239, 248)" />
           <Path
             path={grid}
@@ -150,9 +179,5 @@ export function MNISTInference() {
 const style = StyleSheet.create({
   container: {
     flex: 1,
-  },
-  canvas: {
-    width,
-    height: width * 2,
   },
 });
