+++ packages/webgpu/src/NativeWebGPUModuleWrapper.ts
@@ -1,0 +1,1 @@
+export { default as WebGPUModule } from "./NativeWebGPUModule";
