+++ packages/webgpu/src/WebGPUViewNativeComponent.ts
@@ -2,7 +2,7 @@ import codegenNativeComponent from "react-native/Libraries/Utilities/codegenNati
 import type { ViewProps } from "react-native";
 import type { Int32 } from "react-native/Libraries/Types/CodegenTypes";
 
-interface NativeProps extends ViewProps {
+export interface NativeProps extends ViewProps {
   contextId: Int32;
   transparent: boolean;
 }
