+++ packages/webgpu/src/WebGPUWrapper.ts
@@ -1,0 +1,1 @@
+export { default as WebGPUWrapper } from "./WebGPUViewNativeComponent";
