+++ packages/webgpu/src/main.web.tsx
@@ -1,0 +1,7 @@
+export * from "./Canvas";
+export * from "./Offscreen";
+export * from "./WebGPUWrapper";
+export * from "./hooks";
+export * from "./types";
+
+export { WebGPUModule } from "./NativeWebGPUModuleWrapper";
